+++ src/wifi/model/minstrel-ht-wifi-manager.cc
@@ -158,7 +158,7 @@ MinstrelHtWifiManager::~MinstrelHtWifiManager ()
   NS_LOG_FUNCTION (this);
   if (HasHtSupported ())
     {
-      for (uint32_t i = 0; i < m_numGroups; i++)
+      for (uint8_t i = 0; i < m_numGroups; i++)
         {
           m_minstrelGroups[i].ratesFirstMpduTxTimeTable.clear ();
           m_minstrelGroups[i].ratesTxTimeTable.clear ();
@@ -974,12 +974,12 @@ MinstrelHtWifiManager::DoGetRtsTxVector (WifiRemoteStation *st)
 
       WifiMode lastRate = GetMcsSupported (station, mcsIndex);
       uint64_t lastDataRate = lastRate.GetNonHtReferenceRate ();
-      uint32_t nBasicRates = GetNBasicModes ();
+      uint8_t nBasicRates = GetNBasicModes ();
 
       WifiMode rtsRate;
       bool rateFound = false;
 
-      for (uint32_t i = 0; i < nBasicRates; i++)
+      for (uint8_t i = 0; i < nBasicRates; i++)
         {
           uint64_t rate = GetBasicMode (i).GetDataRate (20);
           if (rate <= lastDataRate)
@@ -1261,7 +1261,7 @@ MinstrelHtWifiManager::UpdateStats (MinstrelHtWifiRemoteStation *station)
   station->m_maxProbRate = GetLowestIndex (station);
 
   /// Update throughput and EWMA for each rate inside each group.
-  for (uint32_t j = 0; j < m_numGroups; j++)
+  for (uint8_t j = 0; j < m_numGroups; j++)
     {
       if (station->m_groupsTable[j].m_supported)
         {
@@ -1272,7 +1272,7 @@ MinstrelHtWifiManager::UpdateStats (MinstrelHtWifiRemoteStation *station)
           station->m_groupsTable[j].m_maxTpRate2 = GetLowestIndex (station, j);
           station->m_groupsTable[j].m_maxProbRate = GetLowestIndex (station, j);
 
-          for (uint32_t i = 0; i < m_numRates; i++)
+          for (uint8_t i = 0; i < m_numRates; i++)
             {
               if (station->m_groupsTable[j].m_ratesTable[i].supported)
                 {
@@ -1513,7 +1513,7 @@ MinstrelHtWifiManager::RateInit (MinstrelHtWifiRemoteStation *station)
   * Initialize groups supported by the receiver.
   */
   NS_LOG_DEBUG ("Supported groups by station:");
-  for (uint32_t groupId = 0; groupId < m_numGroups; groupId++)
+  for (uint8_t groupId = 0; groupId < m_numGroups; groupId++)
     {
       if (m_minstrelGroups[groupId].isSupported)
         {
@@ -1532,13 +1532,13 @@ MinstrelHtWifiManager::RateInit (MinstrelHtWifiRemoteStation *station)
               station->m_groupsTable[groupId].m_index = 0;
 
               station->m_groupsTable[groupId].m_ratesTable = HtMinstrelRate (m_numRates);        ///Create the rate list for the group.
-              for (uint32_t i = 0; i < m_numRates; i++)
+              for (uint8_t i = 0; i < m_numRates; i++)
                 {
                   station->m_groupsTable[groupId].m_ratesTable[i].supported = false;
                 }
 
               // Initialize all modes supported by the remote station that belong to the current group.
-              for (uint32_t i = 0; i < station->m_nModes; i++)
+              for (uint8_t i = 0; i < station->m_nModes; i++)
                 {
                   WifiMode mode = GetMcsSupported (station, i);
 
@@ -1667,7 +1667,7 @@ MinstrelHtWifiManager::InitSampleTable (MinstrelHtWifiRemoteStation *station)
   station->m_col = station->m_index = 0;
 
   //for off-setting to make rates fall between 0 and nModes
-  uint32_t numSampleRates = m_numRates;
+  uint8_t numSampleRates = m_numRates;
 
   uint32_t newIndex;
   for (uint32_t col = 0; col < m_nSampleCol; col++)
@@ -1699,7 +1699,7 @@ MinstrelHtWifiManager::PrintTable (MinstrelHtWifiRemoteStation *station)
 
   station->m_statsFile << "               best   ____________rate__________    ________statistics________    ________last_______    ______sum-of________\n" <<
     " mode guard #  rate  [name   idx airtime  max_tp]  [avg(tp) avg(prob) sd(prob)]  [prob.|retry|suc|att]  [#success | #attempts]\n";
-  for (uint32_t i = 0; i < m_numGroups; i++)
+  for (uint8_t i = 0; i < m_numGroups; i++)
     {
       StatsDump (station, i, station->m_statsFile);
     }
@@ -1714,7 +1714,7 @@ MinstrelHtWifiManager::PrintTable (MinstrelHtWifiRemoteStation *station)
 void
 MinstrelHtWifiManager::StatsDump (MinstrelHtWifiRemoteStation *station, uint32_t groupId, std::ofstream &of)
 {
-  uint32_t numRates = m_numRates;
+  uint8_t numRates = m_numRates;
   McsGroup group = m_minstrelGroups[groupId];
   Time txTime;
   char giMode;
@@ -1726,7 +1726,7 @@ MinstrelHtWifiManager::StatsDump (MinstrelHtWifiRemoteStation *station, uint32_t
     {
       giMode = 'L';
     }
-  for (uint32_t i = 0; i < numRates; i++)
+  for (uint8_t i = 0; i < numRates; i++)
     {
       if (station->m_groupsTable[groupId].m_supported && station->m_groupsTable[groupId].m_ratesTable[i].supported)
         {
@@ -1879,7 +1879,7 @@ MinstrelHtWifiManager::GetVhtDeviceMcsList (void) const
 {
   WifiModeList vhtMcsList;
   Ptr<WifiPhy> phy = GetPhy ();
-  for (uint32_t i = 0; i < phy->GetNMcs (); i++)
+  for (uint8_t i = 0; i < phy->GetNMcs (); i++)
     {
       WifiMode mode = phy->GetMcs (i);
       if (mode.GetModulationClass () == WIFI_MOD_CLASS_VHT)
@@ -1895,7 +1895,7 @@ MinstrelHtWifiManager::GetHtDeviceMcsList (void) const
 {
   WifiModeList htMcsList;
   Ptr<WifiPhy> phy = GetPhy ();
-  for (uint32_t i = 0; i < phy->GetNMcs (); i++)
+  for (uint8_t i = 0; i < phy->GetNMcs (); i++)
     {
       WifiMode mode = phy->GetMcs (i);
       if (mode.GetModulationClass () == WIFI_MOD_CLASS_HT)
